requirements.txt
blinker==1.7.0
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
Flask==3.0.3
Flask-Cors==4.0.0
fonttools==4.51.0
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pandas==2.2.2
patsy==0.5.6
pillow==10.3.0
plotly==5.20.0
plotly-express==0.4.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scipy==1.13.0
six==1.16.0
statsmodels==0.14.2
tenacity==8.2.3
tzdata==2024.1
Werkzeug==3.0.2
wordcloud==1.9.3
